requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
opencv-python>=4.8.0
Pillow>=9.5.0
accelerate>=0.20.0
xformers>=0.0.20
tqdm>=4.64.0
matplotlib>=3.5.0
